♻️ refactor(fmt): invert the dependency -- consumers register into the engine
The engine had the layering backwards. `unxt._src.fmt` imported
`unxt._src.quantity.base` for `AbstractQuantity` and `custom_pdoc_no_kind`,
and `unxt.units` for `AbstractUnit`, then registered `pparts` for both. That
made the engine depend on its own consumers, and the resulting cycle is what
forced `quantity/base.py`, `quantity/mixins.py` and `unitsystems/base.py` to
import it lazily inside methods, each with a `pylint: disable` explaining the
workaround.

Dependencies now point inward. `unxt._src.fmt` imports no `unxt` module at
all -- only stdlib, jax, numpy, wadler-lindig and plum -- and every
domain-specific rendering is registered *into* it:

- `unxt._src.units` registers `pparts` for `AbstractUnit`
- `unxt._src.quantity.base` registers `pparts` and `pspec_fallback` for
  `AbstractQuantity`
- `unxts.linalg` already registered `pparts` for `UnitsMatrix`, and needed no
  change -- which is the evidence that the extension seam was the right shape;
  it was only the built-in types that were wired the wrong way round.

Consequences:

- All three lazy imports become ordinary module-scope imports, and both
  `import-outside-toplevel` disables are deleted.
- `custom_pdoc_no_kind` and `custom_pdoc_noarray` move to the engine. They are
  generic array-to-doc helpers with no quantity knowledge, and the engine
  needs one of them; `__pdoc__` imports them back.
- `_value_str` becomes public `value_str`, and `_bad_spec` becomes `bad_spec`,
  since the quantity registration in `base.py` now calls both. A downstream
  composite that factors units out of its components needs `value_str` for the
  same reason.

No behaviour change: same output, same tests, 10.00/10 pylint.
